+++ src/modules/reviewers.py
@@ -24,16 +24,6 @@ class ReviewerController:
 
     def get_reviewer_by_id(self, reviewer_id: int, db: Session = Depends(get_db)):
         return db.query(Reviewers).filter(Reviewers.id == reviewer_id).first()
-
-    def get_top5_states_by_reviews(self, db: Session = Depends(get_db)):
-        top_states = db.query(Reviewers.state, func.count(Reviews.id).label('total_reviews')) \
-                    .join(Reviews, Reviewers.id == Reviews.reviewer_id) \
-                    .group_by(Reviewers.state) \
-                    .order_by(func.count(Reviews.id).desc()) \
-                    .limit(5) \
-                    .all()
-        top_states_reviews = [{"state": state, "total_reviews": total_reviews} for state, total_reviews in top_states]
-        return top_states_reviews
     
     def get_all_reviewers_by_state(self, state:str):
         try:
